requirements.txt
pillow==8.3.2
numpy==1.21.2
opencv-python==4.5.5.64
torch==2.3.0
pyyaml==6.0
albumentations==1.1.0
matplotlib==3.5.0
streamlit==1.32.0